refactor(console-ui): phase 2 — dedup Stripe payouts into components/payouts/ (F3)
The entire Stripe Connect payouts feature was duplicated across billing and
earnings (~340 byte-identical lines; the country dropdown was even duplicated
twice within earnings). Extracted a shared module:

- PayoutModal, MethodOption, StripeWithdrawModal, WithdrawalsList — verbatim
  shared leaves.
- CountryPicker — the searchable dropdown (now also replaces billing's plain
  <select>, unifying the cosmetic fork the proposal flagged).
- StripePayoutsCard — one parameterized card (title/icon/className/noun + balance
  slot) for both pages.
- useStripePayouts — the shared status/onboard/withdraw state machine, with
  page-supplied post-withdraw reload + optional analytics.

BillingContent 922 -> ~470 lines, EarningsContent 850 -> ~340 lines. Earnings
poll is now visibility-gated and uses the central clientCoordinatorUrl resolver.

Tests: +__tests__/payouts.test.tsx (hook flows + card states). Invariant held:
build green, eslint 0 errors / 130 warnings (down from 137), vitest 344 pass /
4 pre-existing fail.
