+++ console-ui/src/app/billing/BillingContent.tsx
@@ -9,20 +9,12 @@ import {
   fetchUsage,
   createStripeCheckout,
   redeemInviteCode,
-  fetchStripeStatus,
-  startStripeOnboarding,
-  withdrawStripe,
-  fetchStripeWithdrawals,
-  computeStripeFeeUsd,
   type BalanceResponse,
   type UsageEntry,
-  type StripeStatus,
-  type StripeWithdrawal,
 } from "@/lib/api";
 import { trackEvent } from "@/lib/google-analytics";
 import {
   Clock,
-  X,
   Loader2,
   DollarSign,
   TrendingUp,
@@ -30,38 +22,14 @@ import {
   Check,
   CreditCard,
   Building2,
-  Zap,
-  ArrowDownToLine,
 } from "lucide-react";
 import { UsageChart } from "@/components/UsageChart";
-import { STRIPE_CONNECT_COUNTRIES } from "@/lib/stripe-countries";
-
-function Modal({
-  open,
-  onClose,
-  children,
-}: {
-  open: boolean;
-  onClose: () => void;
-  children: React.ReactNode;
-}) {
-  if (!open) return null;
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/40 backdrop-blur-sm">
-      <div className="bg-bg-white border border-border-dim rounded-xl w-full max-w-md mx-2 sm:mx-4 shadow-lg">
-        <div className="flex justify-end p-3">
-          <button
-            onClick={onClose}
-            className="p-1 rounded hover:bg-bg-hover text-text-tertiary"
-          >
-            <X size={16} />
-          </button>
-        </div>
-        {children}
-      </div>
-    </div>
-  );
-}
+import {
+  PayoutModal,
+  StripePayoutsCard,
+  StripeWithdrawModal,
+  useStripePayouts,
+} from "@/components/payouts";
 
 export default function BillingContent() {
   const addToast = useToastStore((s) => s.addToast);
@@ -80,25 +48,6 @@ export default function BillingContent() {
   );
   const [sortAsc, setSortAsc] = useState(false);
 
-  // Stripe Payouts state.
-  const [stripeStatus, setStripeStatus] = useState<StripeStatus | null>(null);
-  const [stripeWithdrawals, setStripeWithdrawals] = useState<StripeWithdrawal[]>([]);
-  const [stripeOnboardLoading, setStripeOnboardLoading] = useState(false);
-  const [withdrawOpen, setWithdrawOpen] = useState(false);
-  const [withdrawAmount, setWithdrawAmount] = useState("10");
-  const [withdrawMethod, setWithdrawMethod] = useState<"standard" | "instant">("standard");
-  const [withdrawLoading, setWithdrawLoading] = useState(false);
-  const [selectedStripeCountry, setSelectedStripeCountry] = useState("");
-
-  // Once a Stripe Express account exists, its country is locked. Pre-select
-  // that country so the user sees what will actually be used, and so a
-  // deliberate change triggers backend creation of a new account.
-  useEffect(() => {
-    if (stripeStatus?.stripe_account_country) {
-      setSelectedStripeCountry(stripeStatus.stripe_account_country);
-    }
-  }, [stripeStatus?.stripe_account_country]);
-
   const loadData = useCallback(async () => {
     setLoading(true);
     try {
@@ -114,44 +63,14 @@ export default function BillingContent() {
     setLoading(false);
   }, [addToast]);
 
-  // Stripe payouts data loads independently — we don't want a misconfigured
-  // coordinator to block the rest of the billing page.
-  const loadStripe = useCallback(async (refresh = false) => {
-    try {
-      const [s, wds] = await Promise.all([
-        fetchStripeStatus(refresh),
-        fetchStripeWithdrawals(20).catch(() => [] as StripeWithdrawal[]),
-      ]);
-      setStripeStatus(s);
-      setStripeWithdrawals(wds);
-    } catch (e) {
-      // Silent — Stripe Payouts is optional infrastructure.
-      console.warn("stripe status fetch failed:", (e as Error).message);
-    }
-  }, []);
+  // Stripe payouts state machine (status/onboard/withdraw) shared with earnings.
+  const payouts = useStripePayouts({ addToast, onAfterWithdraw: loadData });
 
   useEffect(() => {
     loadData();
   }, [loadData]);
 
-  // Refresh Stripe status on mount; if the user just came back from the
-  // Stripe-hosted onboarding flow (return URL adds ?stripe_return=1) we hit
-  // ?refresh=1 so the coordinator pulls the latest snapshot from Stripe
-  // before the webhook arrives.
-  useEffect(() => {
-    const params = typeof window !== "undefined" ? new URLSearchParams(window.location.search) : null;
-    const justReturned = params?.get("stripe_return") === "1";
-    loadStripe(justReturned);
-    if (justReturned) {
-      addToast("Stripe onboarding complete — verifying...", "success");
-      // Strip the query param so a refresh doesn't re-trigger the toast.
-      const url = new URL(window.location.href);
-      url.searchParams.delete("stripe_return");
-      window.history.replaceState({}, "", url.toString());
-    }
-  }, [loadStripe, addToast]);
-
-  // Detect Stripe Checkout success redirect
+  // Detect Stripe Checkout (buy credits) success redirect.
   useEffect(() => {
     if (typeof window === "undefined") return;
     const params = new URLSearchParams(window.location.search);
@@ -163,36 +82,6 @@ export default function BillingContent() {
       window.history.replaceState({}, "", url.toString());
     }
   }, [addToast, loadData]);
-
-  const handleStripeOnboard = async () => {
-    setStripeOnboardLoading(true);
-    try {
-      // We pass the current page (with ?stripe_return=1) as the return URL so
-      // the user lands back on Billing after KYC.
-      const returnURL = typeof window !== "undefined"
-        ? `${window.location.origin}${window.location.pathname}?stripe_return=1`
-        : undefined;
-      const resp = await startStripeOnboarding(returnURL, selectedStripeCountry || undefined);
-      window.location.href = resp.url;
-    } catch (e) {
-      addToast(`Stripe onboarding failed: ${(e as Error).message}`);
-      setStripeOnboardLoading(false);
-    }
-  };
-
-  const handleStripeWithdraw = async () => {
-    setWithdrawLoading(true);
-    try {
-      const resp = await withdrawStripe(withdrawAmount, withdrawMethod);
-      addToast(`Withdrawal submitted — ${resp.eta || "processing"}`, "success");
-      setWithdrawOpen(false);
-      // Reload balance + history.
-      await Promise.all([loadData(), loadStripe(false)]);
-    } catch (e) {
-      addToast(`${(e as Error).message}`);
-    }
-    setWithdrawLoading(false);
-  };
 
   const handleStripeCheckout = async () => {
     setActionLoading(true);
@@ -339,18 +228,18 @@ export default function BillingContent() {
 
           {/* Withdraw to Bank (Stripe Connect Express) */}
           <StripePayoutsCard
-            status={stripeStatus}
-            withdrawals={stripeWithdrawals}
+            status={payouts.status}
+            withdrawals={payouts.withdrawals}
             balanceMicroUsd={balance?.balance_micro_usd ?? 0}
-            onboardLoading={stripeOnboardLoading}
-            selectedCountry={selectedStripeCountry}
-            onCountryChange={setSelectedStripeCountry}
-            onOnboard={handleStripeOnboard}
-            onOpenWithdraw={() => {
-              setWithdrawAmount("10");
-              setWithdrawMethod(stripeStatus?.instant_eligible ? "instant" : "standard");
-              setWithdrawOpen(true);
-            }}
+            onboardLoading={payouts.onboardLoading}
+            selectedCountry={payouts.selectedCountry}
+            onCountryChange={payouts.setSelectedCountry}
+            onOnboard={payouts.onboard}
+            onOpenWithdraw={() => payouts.openWithdraw("10")}
+            title="Withdraw to Bank"
+            icon={<Building2 size={16} className="text-teal" />}
+            noun="credits"
+            className="rounded-2xl border border-border-dim bg-bg-white p-6 shadow-md"
           />
 
           {/* Stats row */}
@@ -476,7 +365,7 @@ export default function BillingContent() {
       </div>
 
       {/* Buy Credits Modal */}
-      <Modal open={buyOpen} onClose={() => setBuyOpen(false)}>
+      <PayoutModal open={buyOpen} onClose={() => setBuyOpen(false)}>
         <div className="px-6 pb-6">
           <h3 className="text-2xl font-semibold text-ink mb-2">
             Buy Credits
@@ -533,390 +422,22 @@ export default function BillingContent() {
             Powered by Stripe. Secure card payment.
           </p>
         </div>
-      </Modal>
+      </PayoutModal>
 
       {/* Stripe Withdraw Modal */}
-      <Modal open={withdrawOpen} onClose={() => !withdrawLoading && setWithdrawOpen(false)}>
+      <PayoutModal open={payouts.withdrawOpen} onClose={() => !payouts.withdrawLoading && payouts.setWithdrawOpen(false)}>
         <StripeWithdrawModal
-          status={stripeStatus}
+          status={payouts.status}
           balanceMicroUsd={balance?.balance_micro_usd ?? 0}
-          amount={withdrawAmount}
-          method={withdrawMethod}
-          loading={withdrawLoading}
-          onAmountChange={setWithdrawAmount}
-          onMethodChange={setWithdrawMethod}
-          onConfirm={handleStripeWithdraw}
-          onCancel={() => setWithdrawOpen(false)}
+          amount={payouts.withdrawAmount}
+          method={payouts.withdrawMethod}
+          loading={payouts.withdrawLoading}
+          onAmountChange={payouts.setWithdrawAmount}
+          onMethodChange={payouts.setWithdrawMethod}
+          onConfirm={payouts.withdraw}
+          onCancel={() => payouts.setWithdrawOpen(false)}
         />
-      </Modal>
+      </PayoutModal>
     </div>
-  );
-}
-
-// --- Stripe Payouts components ---
-
-function StripePayoutsCard({
-  status,
-  withdrawals,
-  balanceMicroUsd,
-  onboardLoading,
-  selectedCountry,
-  onCountryChange,
-  onOnboard,
-  onOpenWithdraw,
-}: {
-  status: StripeStatus | null;
-  withdrawals: StripeWithdrawal[];
-  balanceMicroUsd: number;
-  onboardLoading: boolean;
-  selectedCountry: string;
-  onCountryChange: (country: string) => void;
-  onOnboard: () => void;
-  onOpenWithdraw: () => void;
-}) {
-  // Stripe payouts not configured on this coordinator — hide the card entirely.
-  if (status && !status.configured) return null;
-
-  const ready = status?.status === "ready";
-  const restricted = status?.status === "restricted";
-  const rejected = status?.status === "rejected";
-  const pending = status?.status === "pending";
-  const balanceUsd = balanceMicroUsd / 1_000_000;
-  const minWithdrawUsd = (status?.min_withdraw_micro_usd ?? 1_000_000) / 1_000_000;
-  const canWithdraw = ready && balanceUsd >= minWithdrawUsd;
-
-  return (
-    <div className="rounded-2xl border border-border-dim bg-bg-white p-6 shadow-md">
-      <div className="flex items-center gap-2 mb-4">
-        <Building2 size={16} className="text-teal" />
-        <h3 className="text-sm font-semibold text-text-primary">Withdraw to Bank</h3>
-        {ready && (
-          <span className="ml-auto text-[10px] font-mono uppercase tracking-widest text-teal bg-teal/10 border border-teal/30 rounded px-2 py-0.5">
-            Ready
-          </span>
-        )}
-        {pending && (
-          <span className="ml-auto text-[10px] font-mono uppercase tracking-widest text-gold bg-gold/10 border border-gold/30 rounded px-2 py-0.5">
-            Pending
-          </span>
-        )}
-        {(restricted || rejected) && (
-          <span className="ml-auto text-[10px] font-mono uppercase tracking-widest text-coral bg-coral/10 border border-coral/30 rounded px-2 py-0.5">
-            Action needed
-          </span>
-        )}
-      </div>
-
-      {!status?.has_account ? (
-        <>
-          <p className="text-sm text-text-secondary mb-4 leading-relaxed">
-            Link a bank account or debit card via Stripe to withdraw your credits.
-            Stripe handles identity verification — onboarding takes about 2 minutes.
-          </p>
-          <label className="block text-xs font-mono text-text-tertiary uppercase tracking-wider mb-2">
-            Your country
-          </label>
-          <select
-            value={selectedCountry}
-            onChange={(e) => onCountryChange(e.target.value)}
-            className="w-full mb-4 bg-bg-primary border border-border-dim rounded-lg px-4 py-3 text-sm text-text-primary outline-none transition-colors hover:border-teal/40 focus:border-teal"
-          >
-            <option value="">Select your country</option>
-            {STRIPE_CONNECT_COUNTRIES.map((country) => (
-              <option key={country.code} value={country.code}>
-                {country.name} ({country.code})
-              </option>
-            ))}
-          </select>
-          <button
-            onClick={onOnboard}
-            disabled={onboardLoading || !selectedCountry}
-            className="flex items-center gap-2 px-5 py-2.5 rounded-lg bg-teal border-2 border-ink text-white text-sm font-bold hover:opacity-90 disabled:opacity-50 transition-all"
-          >
-            {onboardLoading ? <Loader2 size={14} className="animate-spin" /> : <Building2 size={14} />}
-            {onboardLoading ? "Redirecting..." : "Link bank via Stripe"}
-          </button>
-          {!selectedCountry && (
-            <p className="text-xs text-text-tertiary mt-2">
-              Select your country to continue. This determines your payout currency and KYC requirements.
-            </p>
-          )}
-        </>
-      ) : ready ? (
-        <>
-          <div className="rounded-lg bg-bg-primary border border-border-dim p-3 mb-4 flex items-center justify-between">
-            <div className="flex items-center gap-2 text-sm text-text-secondary">
-              {status.destination_type === "card" ? (
-                <CreditCard size={14} className="text-teal" />
-              ) : (
-                <Building2 size={14} className="text-teal" />
-              )}
-              <span className="font-mono">
-                {status.destination_type === "card" ? "Debit card" : "Bank"} ••{status.destination_last4}
-              </span>
-              {status.instant_eligible && (
-                <span className="text-[10px] font-mono uppercase text-gold bg-gold/10 border border-gold/30 rounded px-1.5 py-0.5">
-                  Instant
-                </span>
-              )}
-            </div>
-          </div>
-          <button
-            onClick={onOpenWithdraw}
-            disabled={!canWithdraw}
-            className="flex items-center gap-2 px-5 py-2.5 rounded-lg bg-teal border-2 border-ink text-white text-sm font-bold hover:opacity-90 disabled:opacity-50 disabled:cursor-not-allowed transition-all"
-          >
-            <ArrowDownToLine size={14} />
-            Withdraw
-          </button>
-          {!canWithdraw && balanceUsd < minWithdrawUsd && (
-            <p className="text-xs text-text-tertiary mt-2">
-              Minimum withdrawal is ${minWithdrawUsd.toFixed(2)} — your balance is ${balanceUsd.toFixed(2)}.
-            </p>
-          )}
-        </>
-      ) : (
-        <>
-          <p className="text-sm text-text-secondary mb-4 leading-relaxed">
-            Your Stripe account is locked to{" "}
-            <span className="font-medium text-text-primary">
-              {STRIPE_CONNECT_COUNTRIES.find((c) => c.code === status?.stripe_account_country)?.name || status?.stripe_account_country || "your selected country"}
-            </span>
-            . If that is not correct, select your country below and we will create a new account.
-          </p>
-          <label className="block text-xs font-mono text-text-tertiary uppercase tracking-wider mb-2">
-            Country
-          </label>
-          <select
-            value={selectedCountry}
-            onChange={(e) => onCountryChange(e.target.value)}
-            className="w-full mb-4 bg-bg-primary border border-border-dim rounded-lg px-4 py-3 text-sm text-text-primary outline-none transition-colors hover:border-teal/40 focus:border-teal"
-          >
-            <option value="">Select your country</option>
-            {STRIPE_CONNECT_COUNTRIES.map((country) => (
-              <option key={country.code} value={country.code}>
-                {country.name} ({country.code})
-              </option>
-            ))}
-          </select>
-          <button
-            onClick={onOnboard}
-            disabled={onboardLoading || !selectedCountry}
-            className="flex items-center gap-2 px-5 py-2.5 rounded-lg bg-teal border-2 border-ink text-white text-sm font-bold hover:opacity-90 disabled:opacity-50 transition-all"
-          >
-            {onboardLoading ? <Loader2 size={14} className="animate-spin" /> : <Building2 size={14} />}
-            {onboardLoading ? "Redirecting..." : restricted ? "Provide more info" : "Continue setup"}
-          </button>
-        </>
-      )}
-
-      {withdrawals.length > 0 && (
-        <div className="mt-5 pt-5 border-t border-border-subtle">
-          <p className="text-xs font-mono text-text-tertiary uppercase tracking-wider mb-3">
-            Recent withdrawals
-          </p>
-          <div className="space-y-2">
-            {withdrawals.slice(0, 5).map((w) => (
-              <div key={w.id} className="flex items-center justify-between text-sm">
-                <div className="flex items-center gap-2">
-                  {w.status === "paid" ? (
-                    <Check size={12} className="text-teal" />
-                  ) : w.status === "failed" ? (
-                    <X size={12} className="text-coral" />
-                  ) : (
-                    <Clock size={12} className="text-gold" />
-                  )}
-                  <span className="font-mono text-text-secondary">
-                    ${(w.net_micro_usd / 1_000_000).toFixed(2)}
-                  </span>
-                  <span className="text-[10px] font-mono uppercase text-text-tertiary">
-                    {w.method}
-                  </span>
-                </div>
-                <span className={`text-xs font-mono ${
-                  w.status === "paid" ? "text-teal" :
-                  w.status === "failed" ? "text-coral" :
-                  "text-text-tertiary"
-                }`}>
-                  {w.status}
-                  {w.refunded ? " (refunded)" : ""}
-                </span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function StripeWithdrawModal({
-  status,
-  balanceMicroUsd,
-  amount,
-  method,
-  loading,
-  onAmountChange,
-  onMethodChange,
-  onConfirm,
-  onCancel,
-}: {
-  status: StripeStatus | null;
-  balanceMicroUsd: number;
-  amount: string;
-  method: "standard" | "instant";
-  loading: boolean;
-  onAmountChange: (v: string) => void;
-  onMethodChange: (m: "standard" | "instant") => void;
-  onConfirm: () => void;
-  onCancel: () => void;
-}) {
-  const amountNum = parseFloat(amount) || 0;
-  const balanceUsd = balanceMicroUsd / 1_000_000;
-  const minWithdrawUsd = (status?.min_withdraw_micro_usd ?? 1_000_000) / 1_000_000;
-  const instantBps = status?.instant_fee_bps ?? 150;
-  const instantMinUsd = status?.instant_fee_min_usd ?? 0.5;
-  const fee = computeStripeFeeUsd(amountNum, method, instantBps, instantMinUsd);
-  const net = Math.max(0, amountNum - fee);
-
-  const tooSmall = amountNum > 0 && amountNum < minWithdrawUsd;
-  const tooLarge = amountNum > balanceUsd;
-  const valid = amountNum >= minWithdrawUsd && !tooLarge;
-
-  return (
-    <div className="px-6 pb-6">
-      <h3 className="text-2xl font-semibold text-ink mb-2">Withdraw to {status?.destination_type === "card" ? "card" : "bank"}</h3>
-      <p className="text-sm text-text-secondary mb-4">
-        Funds go to {status?.destination_type === "card" ? "your linked card" : "your linked bank account"} ••{status?.destination_last4}.
-      </p>
-
-      <label className="block text-xs font-mono text-text-tertiary uppercase tracking-wider mb-2">
-        Amount (USD)
-      </label>
-      <div className="flex items-center gap-2 mb-3">
-        <span className="text-text-tertiary text-lg">$</span>
-        <input
-          type="number"
-          value={amount}
-          onChange={(e) => onAmountChange(e.target.value)}
-          className="flex-1 bg-bg-primary border border-border-dim rounded-lg px-4 py-3 text-text-primary font-mono text-lg outline-none focus:border-teal transition-colors"
-          min={minWithdrawUsd}
-          max={balanceUsd}
-          step="0.01"
-        />
-      </div>
-      <p className="text-xs text-text-tertiary mb-4">
-        Available: ${balanceUsd.toFixed(2)} · Min: ${minWithdrawUsd.toFixed(2)}
-      </p>
-      {tooSmall && (
-        <p className="text-xs text-coral mb-3">Minimum withdrawal is ${minWithdrawUsd.toFixed(2)}.</p>
-      )}
-      {tooLarge && (
-        <p className="text-xs text-coral mb-3">Insufficient balance.</p>
-      )}
-
-      {/* Method picker */}
-      <label className="block text-xs font-mono text-text-tertiary uppercase tracking-wider mb-2">
-        Speed
-      </label>
-      <div className="grid grid-cols-1 gap-2 mb-4">
-        <MethodOption
-          selected={method === "standard"}
-          onClick={() => onMethodChange("standard")}
-          icon={<Clock size={14} />}
-          label="Standard"
-          eta="1-2 business days"
-          fee="Free"
-        />
-        <MethodOption
-          selected={method === "instant"}
-          onClick={() => status?.instant_eligible && onMethodChange("instant")}
-          disabled={!status?.instant_eligible}
-          icon={<Zap size={14} />}
-          label="Instant"
-          eta="~30 minutes"
-          fee={`${(instantBps / 100).toFixed(2)}% (min $${instantMinUsd.toFixed(2)})`}
-          tooltip={!status?.instant_eligible ? "Link a debit card via Stripe to enable Instant Payouts" : undefined}
-        />
-      </div>
-
-      {/* Fee breakdown */}
-      <div className="rounded-lg bg-bg-primary border border-border-dim p-3 mb-5 text-xs space-y-1">
-        <div className="flex justify-between text-text-tertiary">
-          <span>Withdrawal</span>
-          <span className="font-mono">${amountNum.toFixed(2)}</span>
-        </div>
-        <div className="flex justify-between text-text-tertiary">
-          <span>Fee</span>
-          <span className="font-mono">${fee.toFixed(2)}</span>
-        </div>
-        <div className="flex justify-between text-text-primary font-bold pt-1 border-t border-border-subtle">
-          <span>You receive</span>
-          <span className="font-mono text-teal">${net.toFixed(2)}</span>
-        </div>
-      </div>
-
-      <div className="flex gap-3">
-        <button
-          onClick={onCancel}
-          disabled={loading}
-          className="flex-1 py-3 rounded-lg border-2 border-border-dim text-text-secondary text-sm font-bold hover:bg-bg-hover transition-all"
-        >
-          Cancel
-        </button>
-        <button
-          onClick={onConfirm}
-          disabled={loading || !valid}
-          className="flex-1 py-3 rounded-lg bg-teal border border-border-dim text-white font-bold text-sm hover:opacity-90 disabled:opacity-50 disabled:cursor-not-allowed transition-all flex items-center justify-center gap-2"
-        >
-          {loading && <Loader2 size={14} className="animate-spin" />}
-          {loading ? "Processing..." : `Withdraw $${amountNum.toFixed(2)}`}
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function MethodOption({
-  selected,
-  onClick,
-  disabled,
-  icon,
-  label,
-  eta,
-  fee,
-  tooltip,
-}: {
-  selected: boolean;
-  onClick: () => void;
-  disabled?: boolean;
-  icon: React.ReactNode;
-  label: string;
-  eta: string;
-  fee: string;
-  tooltip?: string;
-}) {
-  return (
-    <button
-      onClick={onClick}
-      disabled={disabled}
-      title={tooltip}
-      className={`flex items-center justify-between gap-3 px-3 py-3 rounded-lg border-2 transition-all text-left ${
-        selected
-          ? "bg-teal/10 border-teal text-teal"
-          : disabled
-          ? "bg-bg-primary border-border-dim text-text-tertiary cursor-not-allowed opacity-60"
-          : "bg-bg-primary border-border-dim text-text-secondary hover:border-teal/30 hover:text-teal"
-      }`}
-    >
-      <div className="flex items-center gap-3">
-        {icon}
-        <div>
-          <div className="text-sm font-semibold">{label}</div>
-          <div className="text-xs font-mono text-text-tertiary">{eta}</div>
-        </div>
-      </div>
-      <div className="text-xs font-mono">{fee}</div>
-    </button>
   );
 }
