+++ console-ui/src/app/providers/earnings/EarningsContent.tsx
@@ -1,18 +1,12 @@
 "use client";
 
-import { useEffect, useState, useCallback, useRef } from "react";
+import { useState, useCallback } from "react";
 import { useAuth } from "@/hooks/useAuth";
 import { trackEvent } from "@/lib/google-analytics";
 import { useToastStore } from "@/hooks/useToast";
-import {
-  fetchStripeStatus,
-  startStripeOnboarding,
-  withdrawStripe,
-  fetchStripeWithdrawals,
-  computeStripeFeeUsd,
-  type StripeStatus,
-  type StripeWithdrawal,
-} from "@/lib/api";
+import { useVisiblePolling } from "@/hooks/useVisiblePolling";
+import { STORAGE_KEYS } from "@/lib/constants";
+import { clientCoordinatorUrl } from "@/lib/coordinator-url";
 import {
   Loader2,
   DollarSign,
@@ -20,17 +14,13 @@ import {
   TrendingUp,
   LogIn,
   ArrowDownToLine,
-  Check,
-  Building2,
-  CreditCard,
-  Clock,
-  Zap,
-  X,
-  ChevronDown,
-  Globe,
-  Search,
 } from "lucide-react";
-import { STRIPE_CONNECT_COUNTRIES } from "@/lib/stripe-countries";
+import {
+  PayoutModal,
+  StripePayoutsCard,
+  StripeWithdrawModal,
+  useStripePayouts,
+} from "@/components/payouts";
 
 interface Earning {
   id: number;
@@ -58,33 +48,6 @@ interface EarningsResponse {
   withdrawable_balance_usd: string;
 }
 
-function Modal({
-  open,
-  onClose,
-  children,
-}: {
-  open: boolean;
-  onClose: () => void;
-  children: React.ReactNode;
-}) {
-  if (!open) return null;
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/40 backdrop-blur-sm">
-      <div className="bg-bg-white border border-border-dim rounded-xl w-full max-w-md mx-2 sm:mx-4 shadow-lg">
-        <div className="flex justify-end p-3">
-          <button
-            onClick={onClose}
-            className="p-1 rounded hover:bg-bg-hover text-text-tertiary"
-          >
-            <X size={16} />
-          </button>
-        </div>
-        {children}
-      </div>
-    </div>
-  );
-}
-
 export default function EarningsContent() {
   const { authenticated, login, getAccessToken } = useAuth();
   const addToast = useToastStore((s) => s.addToast);
@@ -92,52 +55,23 @@ export default function EarningsContent() {
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
 
-  // Stripe Payouts state
-  const [stripeStatus, setStripeStatus] = useState<StripeStatus | null>(null);
-  const [stripeWithdrawals, setStripeWithdrawals] = useState<StripeWithdrawal[]>([]);
-  const [stripeOnboardLoading, setStripeOnboardLoading] = useState(false);
-  const [withdrawOpen, setWithdrawOpen] = useState(false);
-  const [withdrawAmount, setWithdrawAmount] = useState("10");
-  const [withdrawMethod, setWithdrawMethod] = useState<"standard" | "instant">("standard");
-  const [withdrawLoading, setWithdrawLoading] = useState(false);
-  const [selectedCountry, setSelectedCountry] = useState("");
-
-  // Once a Stripe Express account exists, its country is locked. Pre-select
-  // that country so the user sees what will actually be used, and so a
-  // deliberate change triggers backend creation of a new account.
-  useEffect(() => {
-    if (stripeStatus?.stripe_account_country) {
-      setSelectedCountry(stripeStatus.stripe_account_country);
-    }
-  }, [stripeStatus?.stripe_account_country]);
-  const [countryDropdownOpen, setCountryDropdownOpen] = useState(false);
-  const [countryFilter, setCountryFilter] = useState("");
-  const countryDropdownRef = useRef<HTMLDivElement>(null);
-
   const getAuthHeaders = useCallback(async () => {
     const accessToken = await getAccessToken().catch(() => null);
     if (accessToken) {
       return { Authorization: `Bearer ${accessToken}` };
     }
 
-    const apiKey = localStorage.getItem("darkbloom_api_key") || "";
+    const apiKey = localStorage.getItem(STORAGE_KEYS.apiKey) || "";
     return apiKey ? { Authorization: `Bearer ${apiKey}` } : {};
   }, [getAccessToken]);
 
   const fetchEarnings = useCallback(async () => {
     setError(null);
     try {
-      const coordinatorUrl =
-        localStorage.getItem("darkbloom_coordinator_url") ||
-        process.env.NEXT_PUBLIC_COORDINATOR_URL ||
-        "https://api.darkbloom.dev";
       const headers = await getAuthHeaders();
-
       const res = await fetch(
-        `${coordinatorUrl}/v1/provider/account-earnings?limit=100`,
-        {
-          headers,
-        }
+        `${clientCoordinatorUrl()}/v1/provider/account-earnings?limit=100`,
+        { headers }
       );
       if (!res.ok) throw new Error(`HTTP ${res.status}`);
       setData(await res.json());
@@ -148,91 +82,21 @@ export default function EarningsContent() {
     }
   }, [getAuthHeaders]);
 
-  const loadStripe = useCallback(async (refresh = false) => {
-    try {
-      const [s, wds] = await Promise.all([
-        fetchStripeStatus(refresh),
-        fetchStripeWithdrawals(20).catch(() => [] as StripeWithdrawal[]),
-      ]);
-      setStripeStatus(s);
-      setStripeWithdrawals(wds);
-    } catch (e) {
-      console.warn("stripe status fetch failed:", (e as Error).message);
-    }
-  }, []);
+  // Stripe payouts state machine shared with billing.
+  const payouts = useStripePayouts({
+    addToast,
+    enabled: authenticated,
+    onAfterWithdraw: fetchEarnings,
+    onWithdrawStart: (method) =>
+      trackEvent("provider_withdraw_started", { surface: "provider_earnings", method }),
+    onWithdrawSuccess: (method) =>
+      trackEvent("provider_withdraw_succeeded", { surface: "provider_earnings", method }),
+    onWithdrawError: () =>
+      trackEvent("provider_withdraw_failed", { surface: "provider_earnings" }),
+  });
 
-  useEffect(() => {
-    if (!authenticated) {
-      setLoading(false);
-      return;
-    }
-    fetchEarnings();
-    const interval = setInterval(fetchEarnings, 30000);
-    return () => clearInterval(interval);
-  }, [authenticated, fetchEarnings]);
-
-  // Load Stripe status; detect return from Stripe onboarding
-  useEffect(() => {
-    if (!authenticated) return;
-    const params = typeof window !== "undefined" ? new URLSearchParams(window.location.search) : null;
-    const justReturned = params?.get("stripe_return") === "1";
-    loadStripe(justReturned);
-    if (justReturned) {
-      addToast("Stripe onboarding complete — verifying...", "success");
-      const url = new URL(window.location.href);
-      url.searchParams.delete("stripe_return");
-      window.history.replaceState({}, "", url.toString());
-    }
-  }, [authenticated, loadStripe, addToast]);
-
-  useEffect(() => {
-    if (!countryDropdownOpen) return;
-    const handleClick = (e: MouseEvent) => {
-      if (countryDropdownRef.current && !countryDropdownRef.current.contains(e.target as Node)) {
-        setCountryDropdownOpen(false);
-      }
-    };
-    document.addEventListener("mousedown", handleClick);
-    return () => document.removeEventListener("mousedown", handleClick);
-  }, [countryDropdownOpen]);
-
-  const handleStripeOnboard = async () => {
-    setStripeOnboardLoading(true);
-    try {
-      const returnURL = typeof window !== "undefined"
-        ? `${window.location.origin}${window.location.pathname}?stripe_return=1`
-        : undefined;
-      const resp = await startStripeOnboarding(returnURL, selectedCountry || undefined);
-      window.location.href = resp.url;
-    } catch (e) {
-      addToast(`Stripe onboarding failed: ${(e as Error).message}`);
-      setStripeOnboardLoading(false);
-    }
-  };
-
-  const handleStripeWithdraw = async () => {
-    setWithdrawLoading(true);
-    trackEvent("provider_withdraw_started", {
-      surface: "provider_earnings",
-      method: withdrawMethod,
-    });
-    try {
-      const resp = await withdrawStripe(withdrawAmount, withdrawMethod);
-      trackEvent("provider_withdraw_succeeded", {
-        surface: "provider_earnings",
-        method: withdrawMethod,
-      });
-      addToast(`Withdrawal submitted — ${resp.eta || "processing"}`, "success");
-      setWithdrawOpen(false);
-      await Promise.all([fetchEarnings(), loadStripe(false)]);
-    } catch (e) {
-      trackEvent("provider_withdraw_failed", {
-        surface: "provider_earnings",
-      });
-      addToast(`${(e as Error).message}`);
-    }
-    setWithdrawLoading(false);
-  };
+  // Poll earnings only while the tab is visible (perf F6).
+  useVisiblePolling(fetchEarnings, 30_000, authenticated);
 
   if (!authenticated) {
     return (
@@ -281,13 +145,8 @@ export default function EarningsContent() {
   const totalJobs = data?.count || 0;
   const recentCount = data?.recent_count ?? data?.earnings.length ?? 0;
 
-  const ready = stripeStatus?.status === "ready";
-  const restricted = stripeStatus?.status === "restricted";
-  const rejected = stripeStatus?.status === "rejected";
-  const pending = stripeStatus?.status === "pending";
-  const minWithdrawUsd = (stripeStatus?.min_withdraw_micro_usd ?? 1_000_000) / 1_000_000;
+  const minWithdrawUsd = (payouts.status?.min_withdraw_micro_usd ?? 1_000_000) / 1_000_000;
   const availableUsd = withdrawableBalanceMicro / 1_000_000;
-  const canWithdraw = ready && availableUsd >= minWithdrawUsd;
 
   return (
     <div className="max-w-4xl mx-auto p-6 space-y-6">
@@ -330,29 +189,24 @@ export default function EarningsContent() {
       </div>
 
       {/* Withdraw Earnings (Stripe Connect) */}
-      <div className="rounded-xl bg-bg-secondary shadow-sm p-5">
-        <div className="flex items-center gap-2 mb-1">
-          <ArrowDownToLine size={16} className="text-teal" />
-          <h3 className="text-sm font-semibold text-text-primary">Withdraw Earnings</h3>
-          {ready && (
-            <span className="ml-auto text-[10px] font-mono uppercase tracking-widest text-teal bg-teal/10 border border-teal/30 rounded px-2 py-0.5">
-              Ready
-            </span>
-          )}
-          {pending && (
-            <span className="ml-auto text-[10px] font-mono uppercase tracking-widest text-gold bg-gold/10 border border-gold/30 rounded px-2 py-0.5">
-              Pending
-            </span>
-          )}
-          {(restricted || rejected) && (
-            <span className="ml-auto text-[10px] font-mono uppercase tracking-widest text-coral bg-coral/10 border border-coral/30 rounded px-2 py-0.5">
-              Action needed
-            </span>
-          )}
-        </div>
-
+      <StripePayoutsCard
+        status={payouts.status}
+        withdrawals={payouts.withdrawals}
+        balanceMicroUsd={withdrawableBalanceMicro}
+        onboardLoading={payouts.onboardLoading}
+        selectedCountry={payouts.selectedCountry}
+        onCountryChange={payouts.setSelectedCountry}
+        onOnboard={payouts.onboard}
+        onOpenWithdraw={() =>
+          payouts.openWithdraw(availableUsd >= minWithdrawUsd ? availableUsd.toFixed(2) : "10")
+        }
+        title="Withdraw Earnings"
+        icon={<ArrowDownToLine size={16} className="text-teal" />}
+        noun="earnings"
+        className="rounded-xl bg-bg-secondary shadow-sm p-5"
+      >
         {/* Withdrawable earnings display */}
-        <div className="flex items-baseline gap-1 mb-1 mt-3">
+        <div className="flex items-baseline gap-1 mb-1 mt-1">
           <span className="text-3xl font-bold text-text-primary font-mono tracking-tight">
             ${(totalBalance / 1_000_000).toFixed(2)}
           </span>
@@ -364,254 +218,7 @@ export default function EarningsContent() {
             <span>${(creditsBalance / 1_000_000).toFixed(2)} credits (non-withdrawable)</span>
           )}
         </div>
-
-        {!stripeStatus?.has_account ? (
-          <>
-            <p className="text-sm text-text-secondary mb-4 leading-relaxed">
-              Link a bank account or debit card via Stripe to withdraw your earnings.
-              Stripe handles identity verification — onboarding takes about 2 minutes.
-            </p>
-
-            {/* Country picker */}
-            <label className="block text-xs font-mono text-text-tertiary uppercase tracking-wider mb-2">
-              Your country
-            </label>
-            <div className="relative mb-4" ref={countryDropdownRef}>
-              <button
-                type="button"
-                onClick={() => { setCountryDropdownOpen(!countryDropdownOpen); setCountryFilter(""); }}
-                className="w-full flex items-center justify-between gap-2 bg-bg-primary border border-border-dim rounded-lg px-4 py-3 text-sm text-left transition-colors hover:border-teal/40 focus:outline-none focus:border-teal"
-              >
-                {selectedCountry ? (
-                  <span className="flex items-center gap-2 text-text-primary">
-                    <span>{STRIPE_CONNECT_COUNTRIES.find(c => c.code === selectedCountry)?.flag}</span>
-                    <span>{STRIPE_CONNECT_COUNTRIES.find(c => c.code === selectedCountry)?.name}</span>
-                  </span>
-                ) : (
-                  <span className="flex items-center gap-2 text-text-tertiary">
-                    <Globe size={14} />
-                    <span>Select your country</span>
-                  </span>
-                )}
-                <ChevronDown size={14} className={`text-text-tertiary transition-transform ${countryDropdownOpen ? "rotate-180" : ""}`} />
-              </button>
-
-              {countryDropdownOpen && (
-                <div className="absolute z-50 mt-1 w-full bg-bg-white border border-border-dim rounded-xl shadow-lg overflow-hidden">
-                  <div className="p-2 border-b border-border-dim">
-                    <div className="relative">
-                      <Search size={14} className="absolute left-3 top-1/2 -translate-y-1/2 text-text-tertiary" />
-                      <input
-                        type="text"
-                        value={countryFilter}
-                        onChange={(e) => setCountryFilter(e.target.value)}
-                        placeholder="Search countries..."
-                        autoFocus
-                        className="w-full bg-bg-primary border border-border-dim rounded-lg pl-9 pr-3 py-2 text-sm text-text-primary placeholder:text-text-tertiary outline-none focus:border-teal"
-                      />
-                    </div>
-                  </div>
-                  <div className="max-h-64 overflow-y-auto">
-                    {STRIPE_CONNECT_COUNTRIES
-                      .filter(c => {
-                        const q = countryFilter.toLowerCase();
-                        return !q || c.name.toLowerCase().includes(q) || c.code.toLowerCase().includes(q);
-                      })
-                      .map(c => (
-                        <button
-                          key={c.code}
-                          type="button"
-                          onClick={() => { setSelectedCountry(c.code); setCountryDropdownOpen(false); }}
-                          className={`w-full flex items-center gap-3 px-4 py-2.5 text-sm text-left transition-colors ${
-                            selectedCountry === c.code
-                              ? "bg-teal/10 text-teal"
-                              : "text-text-secondary hover:bg-bg-hover"
-                          }`}
-                        >
-                          <span className="text-base">{c.flag}</span>
-                          <span className="flex-1">{c.name}</span>
-                          <span className="text-xs font-mono text-text-tertiary">{c.code}</span>
-                        </button>
-                      ))
-                    }
-                  </div>
-                </div>
-              )}
-            </div>
-
-            <button
-              onClick={handleStripeOnboard}
-              disabled={stripeOnboardLoading || !selectedCountry}
-              className="flex items-center gap-2 px-5 py-2.5 rounded-lg bg-teal border-2 border-ink text-white text-sm font-bold hover:opacity-90 disabled:opacity-50 disabled:cursor-not-allowed transition-all"
-            >
-              {stripeOnboardLoading ? <Loader2 size={14} className="animate-spin" /> : <Building2 size={14} />}
-              {stripeOnboardLoading ? "Redirecting..." : "Link bank via Stripe"}
-            </button>
-            {!selectedCountry && (
-              <p className="text-xs text-text-tertiary mt-2">
-                Select your country to continue. This determines your payout currency and KYC requirements.
-              </p>
-            )}
-          </>
-        ) : ready ? (
-          <>
-            <div className="rounded-lg bg-bg-primary border border-border-dim p-3 mb-4 flex items-center justify-between">
-              <div className="flex items-center gap-2 text-sm text-text-secondary">
-                {stripeStatus.destination_type === "card" ? (
-                  <CreditCard size={14} className="text-teal" />
-                ) : (
-                  <Building2 size={14} className="text-teal" />
-                )}
-                <span className="font-mono">
-                  {stripeStatus.destination_type === "card" ? "Debit card" : "Bank"} ••{stripeStatus.destination_last4}
-                </span>
-                {stripeStatus.instant_eligible && (
-                  <span className="text-[10px] font-mono uppercase text-gold bg-gold/10 border border-gold/30 rounded px-1.5 py-0.5">
-                    Instant
-                  </span>
-                )}
-              </div>
-            </div>
-            <button
-              onClick={() => {
-                setWithdrawAmount(availableUsd >= minWithdrawUsd ? availableUsd.toFixed(2) : "10");
-                setWithdrawMethod(stripeStatus?.instant_eligible ? "instant" : "standard");
-                setWithdrawOpen(true);
-              }}
-              disabled={!canWithdraw}
-              className="flex items-center gap-2 px-5 py-2.5 rounded-lg bg-teal border-2 border-ink text-white text-sm font-bold hover:opacity-90 disabled:opacity-50 disabled:cursor-not-allowed transition-all"
-            >
-              <ArrowDownToLine size={14} />
-              Withdraw
-            </button>
-            {!canWithdraw && availableUsd < minWithdrawUsd && (
-              <p className="text-xs text-text-tertiary mt-2">
-                Minimum withdrawal is ${minWithdrawUsd.toFixed(2)} — your available balance is ${availableUsd.toFixed(2)}.
-              </p>
-            )}
-          </>
-        ) : (
-          <>
-            <p className="text-sm text-text-secondary mb-4 leading-relaxed">
-              Your Stripe account is locked to{" "}
-              <span className="font-medium text-text-primary">
-                {STRIPE_CONNECT_COUNTRIES.find(c => c.code === stripeStatus?.stripe_account_country)?.name || stripeStatus?.stripe_account_country || "your selected country"}
-              </span>
-              . If that is not correct, select your country below and we will create a new account.
-            </p>
-            <label className="block text-xs font-mono text-text-tertiary uppercase tracking-wider mb-2">
-              Country
-            </label>
-            <div className="relative mb-4" ref={countryDropdownRef}>
-              <button
-                type="button"
-                onClick={() => { setCountryDropdownOpen(!countryDropdownOpen); setCountryFilter(""); }}
-                className="w-full flex items-center justify-between gap-2 bg-bg-primary border border-border-dim rounded-lg px-4 py-3 text-sm text-left transition-colors hover:border-teal/40 focus:outline-none focus:border-teal"
-              >
-                {selectedCountry ? (
-                  <span className="flex items-center gap-2 text-text-primary">
-                    <span>{STRIPE_CONNECT_COUNTRIES.find(c => c.code === selectedCountry)?.flag}</span>
-                    <span>{STRIPE_CONNECT_COUNTRIES.find(c => c.code === selectedCountry)?.name}</span>
-                  </span>
-                ) : (
-                  <span className="flex items-center gap-2 text-text-tertiary">
-                    <Globe size={14} />
-                    <span>Select your country</span>
-                  </span>
-                )}
-                <ChevronDown size={14} className={`text-text-tertiary transition-transform ${countryDropdownOpen ? "rotate-180" : ""}`} />
-              </button>
-
-              {countryDropdownOpen && (
-                <div className="absolute z-50 mt-1 w-full bg-bg-white border border-border-dim rounded-xl shadow-lg overflow-hidden">
-                  <div className="p-2 border-b border-border-dim">
-                    <div className="relative">
-                      <Search size={14} className="absolute left-3 top-1/2 -translate-y-1/2 text-text-tertiary" />
-                      <input
-                        type="text"
-                        value={countryFilter}
-                        onChange={(e) => setCountryFilter(e.target.value)}
-                        placeholder="Search countries..."
-                        autoFocus
-                        className="w-full bg-bg-primary border border-border-dim rounded-lg pl-9 pr-3 py-2 text-sm text-text-primary placeholder:text-text-tertiary outline-none focus:border-teal"
-                      />
-                    </div>
-                  </div>
-                  <div className="max-h-64 overflow-y-auto">
-                    {STRIPE_CONNECT_COUNTRIES
-                      .filter(c => {
-                        const q = countryFilter.toLowerCase();
-                        return !q || c.name.toLowerCase().includes(q) || c.code.toLowerCase().includes(q);
-                      })
-                      .map(c => (
-                        <button
-                          key={c.code}
-                          type="button"
-                          onClick={() => { setSelectedCountry(c.code); setCountryDropdownOpen(false); }}
-                          className={`w-full flex items-center gap-3 px-4 py-2.5 text-sm text-left transition-colors ${
-                            selectedCountry === c.code
-                              ? "bg-teal/10 text-teal"
-                              : "text-text-secondary hover:bg-bg-hover"
-                          }`}
-                        >
-                          <span className="text-base">{c.flag}</span>
-                          <span className="flex-1">{c.name}</span>
-                          <span className="text-xs font-mono text-text-tertiary">{c.code}</span>
-                        </button>
-                      ))
-                    }
-                  </div>
-                </div>
-              )}
-            </div>
-            <button
-              onClick={handleStripeOnboard}
-              disabled={stripeOnboardLoading || !selectedCountry}
-              className="flex items-center gap-2 px-5 py-2.5 rounded-lg bg-teal border-2 border-ink text-white text-sm font-bold hover:opacity-90 disabled:opacity-50 disabled:cursor-not-allowed transition-all"
-            >
-              {stripeOnboardLoading ? <Loader2 size={14} className="animate-spin" /> : <Building2 size={14} />}
-              {stripeOnboardLoading ? "Redirecting..." : restricted ? "Provide more info" : "Continue setup"}
-            </button>
-          </>
-        )}
-
-        {stripeWithdrawals.length > 0 && (
-          <div className="mt-5 pt-5 border-t border-border-subtle">
-            <p className="text-xs font-mono text-text-tertiary uppercase tracking-wider mb-3">
-              Recent withdrawals
-            </p>
-            <div className="space-y-2">
-              {stripeWithdrawals.slice(0, 5).map((w) => (
-                <div key={w.id} className="flex items-center justify-between text-sm">
-                  <div className="flex items-center gap-2">
-                    {w.status === "paid" ? (
-                      <Check size={12} className="text-teal" />
-                    ) : w.status === "failed" ? (
-                      <X size={12} className="text-coral" />
-                    ) : (
-                      <Clock size={12} className="text-gold" />
-                    )}
-                    <span className="font-mono text-text-secondary">
-                      ${(w.net_micro_usd / 1_000_000).toFixed(2)}
-                    </span>
-                    <span className="text-[10px] font-mono uppercase text-text-tertiary">
-                      {w.method}
-                    </span>
-                  </div>
-                  <span className={`text-xs font-mono ${
-                    w.status === "paid" ? "text-teal" :
-                    w.status === "failed" ? "text-coral" :
-                    "text-text-tertiary"
-                  }`}>
-                    {w.status}
-                    {w.refunded ? " (refunded)" : ""}
-                  </span>
-                </div>
-              ))}
-            </div>
-          </div>
-        )}
-      </div>
+      </StripePayoutsCard>
 
       {/* Earnings history */}
       <div>
@@ -661,190 +268,19 @@ export default function EarningsContent() {
       </div>
 
       {/* Stripe Withdraw Modal */}
-      <Modal open={withdrawOpen} onClose={() => !withdrawLoading && setWithdrawOpen(false)}>
+      <PayoutModal open={payouts.withdrawOpen} onClose={() => !payouts.withdrawLoading && payouts.setWithdrawOpen(false)}>
         <StripeWithdrawModal
-          status={stripeStatus}
+          status={payouts.status}
           balanceMicroUsd={withdrawableBalanceMicro}
-          amount={withdrawAmount}
-          method={withdrawMethod}
-          loading={withdrawLoading}
-          onAmountChange={setWithdrawAmount}
-          onMethodChange={setWithdrawMethod}
-          onConfirm={handleStripeWithdraw}
-          onCancel={() => setWithdrawOpen(false)}
+          amount={payouts.withdrawAmount}
+          method={payouts.withdrawMethod}
+          loading={payouts.withdrawLoading}
+          onAmountChange={payouts.setWithdrawAmount}
+          onMethodChange={payouts.setWithdrawMethod}
+          onConfirm={payouts.withdraw}
+          onCancel={() => payouts.setWithdrawOpen(false)}
         />
-      </Modal>
+      </PayoutModal>
     </div>
-  );
-}
-
-function StripeWithdrawModal({
-  status,
-  balanceMicroUsd,
-  amount,
-  method,
-  loading,
-  onAmountChange,
-  onMethodChange,
-  onConfirm,
-  onCancel,
-}: {
-  status: StripeStatus | null;
-  balanceMicroUsd: number;
-  amount: string;
-  method: "standard" | "instant";
-  loading: boolean;
-  onAmountChange: (v: string) => void;
-  onMethodChange: (m: "standard" | "instant") => void;
-  onConfirm: () => void;
-  onCancel: () => void;
-}) {
-  const amountNum = parseFloat(amount) || 0;
-  const balanceUsd = balanceMicroUsd / 1_000_000;
-  const minWithdrawUsd = (status?.min_withdraw_micro_usd ?? 1_000_000) / 1_000_000;
-  const instantBps = status?.instant_fee_bps ?? 150;
-  const instantMinUsd = status?.instant_fee_min_usd ?? 0.5;
-  const fee = computeStripeFeeUsd(amountNum, method, instantBps, instantMinUsd);
-  const net = Math.max(0, amountNum - fee);
-
-  const tooSmall = amountNum > 0 && amountNum < minWithdrawUsd;
-  const tooLarge = amountNum > balanceUsd;
-  const valid = amountNum >= minWithdrawUsd && !tooLarge;
-
-  return (
-    <div className="px-6 pb-6">
-      <h3 className="text-2xl font-semibold text-ink mb-2">Withdraw to {status?.destination_type === "card" ? "card" : "bank"}</h3>
-      <p className="text-sm text-text-secondary mb-4">
-        Funds go to {status?.destination_type === "card" ? "your linked card" : "your linked bank account"} ••{status?.destination_last4}.
-      </p>
-
-      <label className="block text-xs font-mono text-text-tertiary uppercase tracking-wider mb-2">
-        Amount (USD)
-      </label>
-      <div className="flex items-center gap-2 mb-3">
-        <span className="text-text-tertiary text-lg">$</span>
-        <input
-          type="number"
-          value={amount}
-          onChange={(e) => onAmountChange(e.target.value)}
-          className="flex-1 bg-bg-primary border border-border-dim rounded-lg px-4 py-3 text-text-primary font-mono text-lg outline-none focus:border-teal transition-colors"
-          min={minWithdrawUsd}
-          max={balanceUsd}
-          step="0.01"
-        />
-      </div>
-      <p className="text-xs text-text-tertiary mb-4">
-        Available: ${balanceUsd.toFixed(2)} · Min: ${minWithdrawUsd.toFixed(2)}
-      </p>
-      {tooSmall && (
-        <p className="text-xs text-coral mb-3">Minimum withdrawal is ${minWithdrawUsd.toFixed(2)}.</p>
-      )}
-      {tooLarge && (
-        <p className="text-xs text-coral mb-3">Insufficient balance.</p>
-      )}
-
-      {/* Method picker */}
-      <label className="block text-xs font-mono text-text-tertiary uppercase tracking-wider mb-2">
-        Speed
-      </label>
-      <div className="grid grid-cols-1 gap-2 mb-4">
-        <MethodOption
-          selected={method === "standard"}
-          onClick={() => onMethodChange("standard")}
-          icon={<Clock size={14} />}
-          label="Standard"
-          eta="1-2 business days"
-          fee="Free"
-        />
-        <MethodOption
-          selected={method === "instant"}
-          onClick={() => status?.instant_eligible && onMethodChange("instant")}
-          disabled={!status?.instant_eligible}
-          icon={<Zap size={14} />}
-          label="Instant"
-          eta="~30 minutes"
-          fee={`${(instantBps / 100).toFixed(2)}% (min $${instantMinUsd.toFixed(2)})`}
-          tooltip={!status?.instant_eligible ? "Link a debit card via Stripe to enable Instant Payouts" : undefined}
-        />
-      </div>
-
-      {/* Fee breakdown */}
-      <div className="rounded-lg bg-bg-primary border border-border-dim p-3 mb-5 text-xs space-y-1">
-        <div className="flex justify-between text-text-tertiary">
-          <span>Withdrawal</span>
-          <span className="font-mono">${amountNum.toFixed(2)}</span>
-        </div>
-        <div className="flex justify-between text-text-tertiary">
-          <span>Fee</span>
-          <span className="font-mono">${fee.toFixed(2)}</span>
-        </div>
-        <div className="flex justify-between text-text-primary font-bold pt-1 border-t border-border-subtle">
-          <span>You receive</span>
-          <span className="font-mono text-teal">${net.toFixed(2)}</span>
-        </div>
-      </div>
-
-      <div className="flex gap-3">
-        <button
-          onClick={onCancel}
-          disabled={loading}
-          className="flex-1 py-3 rounded-lg border-2 border-border-dim text-text-secondary text-sm font-bold hover:bg-bg-hover transition-all"
-        >
-          Cancel
-        </button>
-        <button
-          onClick={onConfirm}
-          disabled={loading || !valid}
-          className="flex-1 py-3 rounded-lg bg-teal border border-border-dim text-white font-bold text-sm hover:opacity-90 disabled:opacity-50 disabled:cursor-not-allowed transition-all flex items-center justify-center gap-2"
-        >
-          {loading && <Loader2 size={14} className="animate-spin" />}
-          {loading ? "Processing..." : `Withdraw $${amountNum.toFixed(2)}`}
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function MethodOption({
-  selected,
-  onClick,
-  disabled,
-  icon,
-  label,
-  eta,
-  fee,
-  tooltip,
-}: {
-  selected: boolean;
-  onClick: () => void;
-  disabled?: boolean;
-  icon: React.ReactNode;
-  label: string;
-  eta: string;
-  fee: string;
-  tooltip?: string;
-}) {
-  return (
-    <button
-      onClick={onClick}
-      disabled={disabled}
-      title={tooltip}
-      className={`flex items-center justify-between gap-3 px-3 py-3 rounded-lg border-2 transition-all text-left ${
-        selected
-          ? "bg-teal/10 border-teal text-teal"
-          : disabled
-          ? "bg-bg-primary border-border-dim text-text-tertiary cursor-not-allowed opacity-60"
-          : "bg-bg-primary border-border-dim text-text-secondary hover:border-teal/30 hover:text-teal"
-      }`}
-    >
-      <div className="flex items-center gap-3">
-        {icon}
-        <div>
-          <div className="text-sm font-semibold">{label}</div>
-          <div className="text-xs font-mono text-text-tertiary">{eta}</div>
-        </div>
-      </div>
-      <div className="text-xs font-mono">{fee}</div>
-    </button>
   );
 }
