+++ console-ui/src/components/payouts/index.ts
@@ -1,0 +1,10 @@
+// Shared Stripe Connect payouts feature. Used by both the billing page and the
+// provider-earnings page, which previously each carried a byte-identical copy
+// of these components + state machine (proposal F3).
+export { PayoutModal } from "./PayoutModal";
+export { MethodOption } from "./MethodOption";
+export { StripeWithdrawModal } from "./StripeWithdrawModal";
+export { CountryPicker } from "./CountryPicker";
+export { WithdrawalsList } from "./WithdrawalsList";
+export { StripePayoutsCard } from "./StripePayoutsCard";
+export { useStripePayouts, type UseStripePayouts } from "./useStripePayouts";
